// src/common/common.types.ts
export const EnvironmentVars = {
  JWT_SECRET: 'JWT_SECRET',
  LOG_QUERIES: 'LOG_QUERIES',
  LOG_REQUESTS: 'LOG_REQUESTS',
  AUTH_COOKIE_SECRET: 'AUTH_COOKIE_SECRET',
  DATABASE_URL: 'DATABASE_URL',
  SMS_ENABLED: 'SMS_ENABLED',
  SMS_LOGIN_ID: 'SMS_LOGIN_ID',
  SMS_PASSWORD: 'SMS_PASSWORD',
  SMS_METHOD: 'SMS_METHOD',
  NOTIFICATION_SERVER_URL: 'NOTIFICATION_SERVER_URL',
  CALL_SERVER_URL: 'CALL_SERVER_URL',
  EXTERNAL_SERVICES_URL: 'EXTERNAL_SERVICES_URL',
  SEND_NOTIFICATIONS: 'SEND_NOTIFICATIONS',
  TWILIO_ACCOUNT_SID: 'TWILIO_ACCOUNT_SID',
  TWILIO_AUTH_TOKEN: 'TWILIO_AUTH_TOKEN',
  TWILIO_PHONE_NUMBER: 'TWILIO_PHONE_NUMBER',
  SEND_CALL_DATA: 'SEND_CALL_DATA',
  SMTP_MAIL_FROM: 'SMTP_MAIL_FROM',
  SMTP_SECRET: 'SMTP_SECRET',
} as const;

export type EnvironmentVars = keyof typeof EnvironmentVars;